write first test, separate out server/app for testing purposes
