+++ backend/config/db.js
@@ -4,13 +4,12 @@ const logger = require('../utils/logger')
 
 logger.info('connecting to', config.MONGODB_URI)
 
-const connectDB = async () => {
-  await mongoose.connect(config.MONGODB_URI)
-    .then(() => {
-      logger.info('connected to MongoDB')
-    })
-    .catch((error) => {
-      logger.error('error connecting to MongoDB:', error.message)
-    })
+const connectDB = () => {
+  try {
+    mongoose.connect(config.MONGODB_URI)
+    logger.info('connected to MongoDB')
+  } catch(error) {
+    logger.error('error connecting to MongoDB:', error.message)
+  }
 }
 module.exports = connectDB
